requirements.txt
langchain
langchain-google-genai
easyocr
fastapi
Pillow
numpy 
opencv-python-headless
uvicorn
